Particle Interactivity Separation + Better documentation
