+++ particles_flutter/lib/interactions.dart
@@ -1,0 +1,1 @@
+export 'src/utils/interactivity.dart';
